Fix to load older kicad_pcb files
The try catch block did not work for deciding which values to load into
the via class. So it is now handled by a if clause that checks if the
version of the file is smaller than the version where the new fields
where introduced. If it is so, we use sensible default values for the
vias.